+++ apps/wallet/src/ui/app/staking/validators/SelectValidatorCard.tsx
