+++ apps/wallet/src/ui/app/staking/validators/ValidatorListItem.tsx
